Add files via upload
In many research projects (for example, in hardware projects), to perform a specific task (for example, changing the multiplication operation, obtaining the maximum, and implementing different functions based on a new hardware), it is necessary to perform that operation in the Forward functions of the layers. different (for example, changing the desired operation in the layer, based on Stochastic calculations), to do this, we must implement the desired layer as a Custom layer so that we can change its Forward function based on the desired task. Therefore, after writing the Forward function of Convolution layers (instead of nn.conv2d), Maxpooling (instead of nn.maxpool2d), FullyConnected (instead of nn.linear), RELU (instead of nn.RELU) in the network, Train and Test network has also been done.